restructured audio code
