--- Audio_test/Audio_test.py
@@ -1,11 +1,0 @@
-from audio.recogniton import LanguageRecognition
-
-audio_file = "Audio.flac"
-
-result = LanguageRecognition(audio_file)  
-
-
-assert result is not None, "Audio processing failed!"
-
-
-print(result)
