+++ test_code_audio/__pycache__/speech2text.cpython-312.pyc
