+++ test_code_audio/test_audioTotext.py
@@ -1,5 +1,5 @@
 # Example usage:
-from speech2text import AudioToText
+from audio.speech2text import AudioToText
 
 att = AudioToText("Your-API-Key")
 output = att.SpeechRecognition("Abhinav/test_audio.wav")
